hw3/requirements.txt
functions-framework==3.10.1
google-cloud-storage==3.9.0
google-cloud-logging==3.13.0
google-cloud-pubsub==2.35.0
